+++ CHANGELOG.md
@@ -125,6 +125,8 @@
 * `BashWrapper`: For int min/max checking: use native bash functionality so there is no dependency to `bc`.
   For double min/max checking: add fallback code to use `awk` in case `bc` is not present on the system (most likely to happen when running tests in a docker container).
 
+* `viash ns list/viash config view`: Allow viewing the post-processed argument groups by passing the `--parse_argument_groups` parameter.
+
 ## TESTING
 
 * `ConfigMod`: Added unit tests for condition config mods.